Fixed security/permission flaws, removed Resolved tab and redirect to My tickets after ticket submission
Based on my responsibilities as directed in our first document, I
started looking at the permissions of roles and what each user should be
able to view.  I found several security flaws where you could do URL
hacking and have access to just about any page in the tracker,
regardless of role. So I put in fixes for those flaws. Then part of my
assignment was to remove the Resolved tab (just remove it) which I have
done but we need to still implement view functionality for resolved
issues in the Views tab.  Finally, it was sort of annoying that as a
student, when submitting a new ticket, it always took you to the "Thank
you for contributing..." page, and then when you clicked Continue, it
redirected you to a page you did not have access to (?).  So I changed
it to skip the "Thank you..." page and go straight to My Tickets after
submitting a ticket.