Updated tabs to reflect user needs
Updated the tabs so that a student (that cannot be assigned to resolve
tickets) no longer has the my work tab. Also made it so that those with
elevated privileges are not shown the Report Issue tab since they will
not likely be reporting issue. Also made the default redirect to the
search page in connection with the upcoming browse/search merge changes.