Merged search and ticket lists together
Made it so that search results are displayed under the search options.
At this point, I was unable to merge My Tickets, My Work and Browse into
a single tab, but I got it down to just Search and My Work sub menus in
the Views tab. There actually was no search logic implemented for the My
Work tab, so I left it because I could not figure out how to integrate
with everything.  Resolved permission stuff with search as well and
fixed an existing bug on My Work page for developer roles (non-manager).